+++ Samples~/Prototype/Architecture/Modules/Sample/ADT/ISampleInterface.cs
@@ -1,0 +1,9 @@
+﻿using EcsRx.Unity.Framework;
+
+namespace BCKWorks.Prototype.Modules.Sample
+{
+    public interface ISampleInterface : IModule
+    {
+        void Print();
+    }
+}