+++ Samples~/Prototype/Architecture/Modules/Sample/IMPL/SampleInterface.cs
@@ -1,0 +1,28 @@
+﻿using EcsRx.Unity.Framework;
+using System.Collections;
+using UnityEngine;
+
+namespace BCKWorks.Prototype.Modules.Sample
+{
+    public class SampleInterface : ISampleInterface
+    {
+        public bool Ready { get; set; }
+
+        public IEnumerator Initialize()
+        {
+            yield return null;
+
+            Ready = true;
+        }
+
+        public void Shutdown()
+        {
+            Ready = false;
+        }
+
+        public void Print()
+        {
+            Debug.Log("Print Sample Interface in Sample Module");
+        }
+    }
+}